+++ hurricane/include/hurricane/util/mat.h
@@ -2,10 +2,10 @@
 #ifndef HC_MAT_H
 #define HC_MAT_H
 
-void hc_mat4x4_identity(double out[4][4]);
-void hc_mat4x4_print(const double mat[4][4]);
-void hc_mat4x4_mul(double a[4][4], double b[4][4], double out[4][4]);
-void hc_mat4x4_mul_vec(double m[4][4], double v[3], double out[3]);
+void hc_mat4x4_identity(float out[4][4]);
+void hc_mat4x4_print(const float mat[4][4]);
+void hc_mat4x4_mul(float a[4][4], float b[4][4], float out[4][4]);
+void hc_mat4x4_mul_vec(float m[4][4], float v[3], float out[3]);
 
 #endif
 */
